[mk/configure] Be more robust to really dirty work directories
It can happen that a work directorie gets really dirty wrt configure cookies,
e.g. when a user is re-configuring a previously built package and hitting
ctrl-c during the re-configuration (yeah, users are cruel). In this case, the
built cookie will exist but not the configure cookie, and if dependencies
also happen to get outdated (because e.g. a git pull changed the package),
gmake will do things in the wrong order.

So make sure that the configure cookie rebuilding is delayed until dependencies
are known to be correct. This is the same trick that is already used for the
depends vs. boostrap-depends cookie.