[mk/depends] Fix brace expansion in prefix in prefixsearch.sh
The expansion of "a{x,y} b{x,y}" is not the same as "a{x,y}b{x,y}", because of
IFS. The first should expand to "ax ay bx by", while the second should expand
to "axbx axby aybx ayby".

So make sure to split words at IFS before passing them to the brace expansion
routine.