[mk/configure] Properly configure cmake's idea of the "make" program
If not set, cmake picks up the first make program in the PATH, which may not be
what we are using (e.g. /usr/pkg/bin/gmake vs. /usr/bin/make ...).

This does not influence cmake itself, but individual packages may want to use
this setting for doing fancy stuff.