[many LAAS/JRL packages] Fix install target hardcoding the "make" program
This fixes the "make" program used to build the doxygen, sphinx or man
documentation of the LAAS/JRL packages using the jrl-cmakemodule cmake
macros. Instead of hardcoding "make" (without any path), use CMAKE_MAKE_PROGRAM
which exists for that purpose. See
https://github.com/jrl-umi3218/jrl-cmakemodules/issues/34

This issue was revealed by the indirect fix of the MAKE_JOBS variable
definition in commit:ce433c9. It happens that NetBSD "make" does not accept
"-j4" as GNUmake does but wants "-j 4" instead. Since MAKE_JOBS was
not properly defined until commit:ce433c9 it was working only by accident.

Packages that are fixed are:
 - graphics/rcpdf
 - graphics/rcpdf-interface
 - interfaces/abstract-git-task
 - interfaces/abstract-robot-dynamics
 - interfaces/jrl-mal
 - math/jrl-dynamics
 - middleware/hpp-template-corba
 - motion/dynamic-graph
 - motion/py-dynamic-graph
 - motion/py-sot-tools
 - motion/sot-core
 - motion/sot-dynamic
 - motion/sot-pattern-generator
 - optimization/roboptim-core
 - optimization/roboptim-trajectory
 - path/hpp-environment-data
 - path/hpp-gik
 - path/hpp-util
 - path/jrl-walkgen