[mk/update] Improve error message when packages have been moved
Instead of suggesting to just remove the package being updated (which works
perfectly, since a make update is about to remove it anyway), be more precise
and suggest to remove only the moved packages.

This has exactly the same result, but it might be less confusing because
people will more likely understand that the faulty package is not the one being
updated but the depending ones that have been moved.