[mk/install] Fix the 'replace' targets for checkouts
Make sure to invalidate the installed state of a package when the 'replace'
target is invoked, by forcing the install cookie to be outdated (as for the
install target itself).

The fixes the 'replace' target with checkouts (and also with up-to-date already
installed packages), because 'replace' relies on install to perform the real
work of uninstalling/reinstalling.