[mk/update] Improve error message when a resumed `make update` fails
`make update` saves the list of packages to be updated before deleting
them. This state persists until all packages are resinstalled or a `make clean
confirm` cancels it. Since this state is a bit hidden, the fact that `make
update` sometimes tries to reinstall other packages might be confusing. Thus, a
more verbose message will now be printed after a failure of a `make update`
involving other packages than the current one.