[mk/depends] Improve sysdep prefix search when a version is not found in a file
commit:ac106e9 allows a sysdep file to be matched whenever the version check
returns no information but exits successfully. The idea was that for some files
the version check may not be mandatory and would be enforced by other files in
the SYSTEM_SEARCH spec.

This strategy fails for tcl on Ubuntu-14.04 where the first tclConfig.sh that
is found is actually a debian shell script including the real file, so the
version check cannot be used on it. 'prefixsearch' returns it without checking
the version and the wrong tcl version may be used in the end.

So this commit updates 'prefixsearch' so that it remembers any file that did
not lead to a proper version information, but continues searching. The file
with no version is used only as a last resort.