[mk/platform] Fix Debian.mk for arm architectures
For the value of SYSLIBDIR, use 'dpkg-architecture' as a fallback if the debian
installation contains multiple architectures.

Also update opsys.sh for arm and make it return 'arm' for all arm variants.

Patch from matthieu@, with updates by me to avoid a strong dependency on
dpkg-architecture in the simple (and most common) case where there is only one
architecture installed.