[mk/sysdep] Fix python packages using distutils
For some reason, the arguments passed to "setup.py install" have always been
wrong. PYSETUPOPTARGS has never been used (drop it). Also, the default setup.py
behaviour is to compile .py files, which is done in a PLIST hook anyway but
raises some issues with python3.2 (not putting the files in the proper
__pycache__ path). So make sure to pass --no-compile. Sigh!