[devel/boost-libs] Make depend.mk more robust
Always search for a regular boost shared library in the SYSTEM_SEARCH
variable, in addition to the versioned one. The latter is needed to get the
boost version, but may be present even when the debian -dev package is not
installed. The regular .so file ensures that one can actually link against
boost.