[motion/sot-dynamic] Release 2.9.2
robotpkg :
- remove patch-aa which has been inserted in cmake module

Release v2.9.2

Split kine_romeo in two scripts
[cmake] Synchronize
Fully support the case of non planar feet case in ZMP from forces.
Correct computations of the X component of the ZMP
Remove the None members from the definition of a humanoid robot.
Correct typo.
Add leftGripper and rightGripper, aliases for the hands.
Correct the creation of the hand frames.
[cmake] Synchronize