[math/eigen3] Disable doc option by default
For some reason, doxygen-1.7 segfaults on this doc. Since this is a rather old
version, there is no point in trying to fix this. But doxygen-1.7 is the only
available version on the OpenNao VMs, so disabling the doc option will make
this package work out of the box on these machines.