[graphics/grh2] Fix install on recent ubuntu and fedora 21
On ubuntu >=14.10, tcl is compiled with gcc -fvisibility=hidded, and grh2
inherit this flag because it uses tclConfig.sh to find the compiler/flags.
So by default, the package entry point was hidden and could not be loaded. Fix
this by adding an explicit attribute visibility=default on the Grh_Init()
function.

Also, make sure to not call Tk function if Tk cannot be loaded (because of a
missing DISPLAY). Tk functions tend to segfault in this context, and this
breaks the package index building. So quit early if Tk cannot be loaded.