[robots/mana-emergency] Fix configure on fedora 21
Let the configure script of the package find either libftdi or libftdi1,
because only the latter is present on Fedora21. Also export the right package
name in the installed .pc, as found by the configure script.

While here, add the dependency on libusb, and export a dependency on libftdi in
the depend.mk, since it is required in the public interface.

Bump PKGREVISION.