[supervision/matlab-genomix] Import matlab-genomix-1.0
matlab-genomix provides a MATLAB package that interacts with the genomix HTTP
server and can control GenoM3 components. It can also be used with the rosix
server to control ros nodes.

The package implements interactive prompting for arguments of
services. Services can be invoked synchronously or asynchronously and callbacks
can be triggered whenever services complete.