[motion/nhfc-genom3] Import nhfc-genom3-1.0
Near-hovering flight controller for rotorcrafts.

This implements a simple stabilization controller based on this publication:
R. Spica, Robuffo Giordano, P., Ryll, M., Bülthoff, H. H., and Franchi, A., “An
Open-Source Hardware/Software Architecture for Quadrotor UAVs”, in 2nd IFAC
Workshop on Research, Education and Development of Unmanned Aerial Systems.

This is a very preliminary release. At the moment this component contains lots
of debugging code. The interface is not stabilized either, and this requires a
motion capture component (hardcoded).