[mk] Fix for dash >= 0.5.8 (found in ubuntu 16.04)
Between 0.5.7 and 0.5.8, dash starts to remove everything from the environment
that does not look like a proper shell variable name:
http://git.kernel.org/cgit/utils/dash/dash.git/commit/?id=46d3c1a614f11f0d40a7e73376359618ff07abcd

And they're not willing to change that:
http://www.mail-archive.com/dash@vger.kernel.org/msg01147.html

So robotpkg must filter out those bad characters from such
variables. Fortunately, this affects only the exported variables that pass down
the required options of packages that are found by the dependency resolution.