[mk] Drop binary packages links in PACKAGES/CATEGORY, provide PACKAGE_SUBDIRS
PACKAGES/CATEGORY have never been used so far (and pkgsrc is on the way to
dropping them as well). Binary packages are now found in the All/ subdirectory
only.

In addition, a new user-settable variable PACKAGES_SUBDIRS is defined, as well
as a corresponding PACKAGES.<subdir> list of package names to put in those
subdirs. This may be used to distribute restricted packages in separate
repositories.

A 'pub' subdirectory is also defined, where packages with no NO_PUBLIC_BIN,
i.e. freely redistributable packages, are linked.