[optimization/acado] Import current acado-1.2.2beta
ACADO Toolkit is a software environment and algorithm collection for automatic
control and dynamic optimization. It provides a general framework for using a
great variety of algorithms for direct optimal control, including model
predictive control, state and parameter estimation and robust optimization.

This package has no release since 2 years and a lot of commits, so the
distribution file is created from the stable branch of the git repository.