[middleware/ros-genmsg] Disable new languages in ros-kinetic
Because ros-kinetic introduces a new 'nodejs' message generation package, it
must be explicitly disabled so that the PLIST of packages is predictable.

If the 'nodejs' messages are ever needed, a more flexible scheme will have to
be introduced.

While here, add the genmsg/__init__.py file in the SYSTEM_SEARCH of this
package so that PYTHONPATH is correctly computed by robotpkg.