[devel/ros-catkin] Manage CATKIN_MARKER_FILE (${PREFIX}/.catkin)
Some ros shell/python script use a $CATKIN_MARKER_FILE to locate packages.
It used to be created by this package, since it is a required dependency for
ros packages. However, this did not work when using a system installation
of ros-catkin.

This commit adds a INSTALL/DEINSTALL pair of script that manage the
creation of the marker file for any package depending on catkin.
The packages using the marker file are tracked within the file itself (since
its content is otherwise not used) and it is deleted when the last package is
removed.

Thanks to Olivier Stasse for the deep analysis of why this file is required and
how it is used by ros scripts.

Bump PKGREVISION.