[simulation/openhrp3] (Re)Import openhrp3-3.1.9
OpenHRP3 was dropped in commit:64217d2 due to the lack of users and complexity
of the package maintenance. This commit re-import the latest version as there
are again users of this, and the newer version are easier to package.

Default to a ros system package on Ubuntu (although this package is unrelated
to ros, system packages are provided in the ros repository).