[robots/hrpsys-base] (Re)Import hrpsys-base
hrpsys-jrl was dropped in commit:089a297 due to the lack of users and
complexity of the package maintenance. This commit re-import the latest version
as there are again users of this, and the newer version are easier to package.
In the meantime, it's been renamed to hrpsys-base.

Default to a ros system package on Ubuntu (although this package is unrelated
to ros, system packages are provided in the ros repository).