[graphics/assimp] Force archive extraction as text
The assimp zip archive has no 'text' label on files, so unzip -a does not
convert CR/LF to LF. The BSD unzip -a is based on a heuristic instead, and does
the conversion. This is an issue for patch-aa that won't apply properly
depending on the unzip program used.

Using unzip -aa makes both versions work the same.