[mk/sysdep] Fix gazebo SYSTEM_SEARCH
Don't put an empty file in SYSTEM_SEARCH.gazebo to make it optional: it
actually matches the $prefix directory, which leads to unexpected results when
building .deb binary package from this (because of dpkg-query -S $prefix).

In this case, the gazebo-config.cmake can actually be found for all versions of
gazebo, in different locations, though. Explicitly put the possibles locations.