[mk] Honour PKG_PATH in pkg-tarup
In pkg-tarup, do not build a binary package file for a dependency if one is
already available in PKG_PATH. This is the same behaviour as before, except
that this honours PKG_PATH to determine if a binary package for a dependency
is already available or not.

The goal of this change is to allow building binary packages for a subset of
robotpkg (typically: wip/) in a different repository than the main one.