[mk] Fix binary package up-to-date check (fallout of commit:9395d3a)
"robotpkg_admin findbest" outputs a URL. Unfortunately, the up-to-date check
for binary package compares dependency file timestamps with
`test -nt`, which does not works with URLs.

The easiest is to pipe the result of "robotpkg_admin findbest" to a sed script
that removes leading "file://" and maps %7e back to ~ (fortunately, ~ is the
only character to deal with, because of the restrictive character set allowed
in package name and options). This will not work for http URLs, but in practice
the bulk builds do not fetch dependencies remotely (and this was not possible
before, so no regression).

While here, also unscramble the package name in the logs: better read ~ instead
of %7e when analyzing the logs ...