[mk] Undo binary package publication in case of failure in the package target
The 'package' target writes outside $WRKDIR. After building the binary package
file itself, a few admistrative tasks must be performed to update the binary
package repository and invoke additional targets like building a debian package
from the binary package.

In case of error in the middle of these administrative tasks, it is safer to
rollback all modifications that were done to the packages repository, so that
it stays in a consistent state.

This is done like for the `install` target, by using a recursive make
invocation, tracking any error and invoking the rollback target if necessary.

As a side effect, a new "depackage" toplevel target is added, which is
available to the users to manually remove a package from the repository.