[mk] Force the bulk target to update uncomplete binary package repository
If, for whatever reason, the packaging of a file and repository updates fails
before completion (e.g. because the building machine is powered down), even
with commit:25f649f the package repository may be left in an inconsistent
state.

To prevent further bulk builds to think that the build was successfull, a
marker file is left along with the binary package file, and removed only at the
very end of the packaging phase. If the file is present when the bulk target
starts, this indicates a previous failure that could not be recovered and the
packaging will be redone.

This should address issue #151