[mk] Add PKGSET_NULLGLOB option, (dis)allowing empty directives in pkg sets
It was initially (and still is) a feature to allow, in a in package set
definition, a package pattern expanding to the empty list. However, this can in
practice hide a typo which may be hard to track down if the package set is
otherwise not empty.

This commit adds a new directive PKGSET_NULLGLOB, which can be specialized per
package set with PKGSET_NULLGLOB.<set-name>, that controls whether a set should
allow patterns exanding to an empty list.

PKGSET_NULLGLOB defaults to no, thus now forbidding empty expansions by
default. It can be turned back to 'yes' to restore the previous behaviour.