[mk] Fix displayed % done in tsort-set for sets that are "strict"
Strict sets skip packages not explicitely listed in their configuration. But
they still process packages dependencies to keep the build order topologically
sorted. The packages outside the strict set are just skipped.

The counter of processed packages must be incremented no matter if they are
skipped or not, otherwise the displayed "% done" is incorrect. This could be
seen on bulk builds for the wip/ directory.