[mk] Fix a corner case issue in dependency constraint reduction
When dealing with multiple constraints on a package (e.g x>n x>m), it could
happen that the lower/upper bounds on the constraint would be incorrectly
computed if the different version numbers would have a different number of
comonents AND the other side of the bound would be longer as well
(e.g: x>=1<3.0 x>=2.1). This was in practice not triggered by any existing
package.

Fix that by making sure to reset the array encoding the min/max constraint
before updating it.