[graphics/flann] Fix build with cmake-3.11
Amend patch-aa fixing the CMakeList : creating a library with no source seem to
not be supported anymore. But this is in practice not needed, this was only
done for the "linux" target, while other target were building libraries
normally. I really don't get the reason to handle linux differently.

Bump PKGREVISION.