[mk] Clean the inter-dependencies of the checksum-vars.mk and checksum.mk files
Like for other *-vars.mk files, checksum-vars.mk should be includable without
triggering the {bootstrap-,}dependencies associated to the actual checksum
targets. This commit leaves only the variable definitions in checksum-vars.mk,
moving the rest to checksum.mk

This way, the checkout target does not artificially inherit from the 'digest'
dependency which is useless when working with a checked out package.