[mk] Compute _ROBOTPKG_NOW in the rule that setups the 'trusted' environment
_ROBOTPKG_NOW is the current date, formatted for computing the "patch level"
version suffix appended for checkout packages. This has to remain constant
during all phases of one make invocation.

The best place to compute this variable is here, instead of in utils.mk,
because some sub-make in robotpkg sometimes resart themselves after dependency
resolution, leading to a new value for the _ROBOTPKG_NOW variable.