[mk] Define PKGTAG for alternative specified via a PKGREQD
When using a PKGREQD to force a specific alternative (e.g. PKGREQD=py36-foo) on
the command line, it can happen that the alternative is invalid. In this case,
an error is reported but the PKGTAG is then defined to the default value
(e.g. py- instead of py36-), which is misleading.

This commit forces the PKGTAG to be consistent with what PKGREQD specifies.