[mk/sysdep] Fix sphinx version detection
Recent sphinx have grown a proper --version option. At the same time, they
stopped reporting the version with an unknown option. Using --version to get
the version will work in both cases, so use that.

Also better filter the output so that the python version is not included by
mistake.