[graphics/ros-urdf] Import 1.13.1 "melodic"
This package contains a C++ parser for the Unified Robot Description
Format (URDF), which is an XML format for representing a robot model.
The code API of the parser has been through our review process and will remain
backwards compatible in future releases.

It is a split of the deprecated ros-robot-model since the melodic release.