[image/opencv2] Update to 2.4.13.6
The complete ChangeLogs are available online:
https://github.com/opencv/opencv/releases

This fixes the support for recent ffmpeg libraries and allows building this
package on recent distributions. The related local patches are not required
anymore. A new patch-aj is needed to support ffmpeg4 libraries, though.

The "doc" option is removed and the doc is not built (this is a mess of
dependencies, including python, sphix, doxygen ... and nobody cares about
doc).

Also remove the "gtk" and "camera" options and enable them
always. Nowadays, a dependency on gtk is almost harmless and V4L is in Linux
anyway. The V4L support is enabled only on Linux, though, since other OS don't
support it well (opencv3 support is better and V4L2 is enabled).