[mk/bulk] Remove erroneous warnings when a package dependency is not available
It can happen that `pkg_add` fails for a dependency of a bulked package. In
this case, the "not availble" error is reported. However, `pkg_add` may have
left unwanted dependencies, that (correctly) won't be removed by `make
deinstall`. Then, warnings will be shown for the package since it does not
leave a clean state after deinstall.

To solve this, make installed dependency as manually installed, with `pkg_add`
instead of `pkg_add -A`. Then, if a pkg_add fails, remove any automatically
installed dependencies that are not required by any manually installed
package. They are necessarily the result of the previously failed pkg_add.