[mk/sysdep] Let openscenegraph-qt dependency check the Qt version
openscenegraph-qt is built with Qt: make sure to use the appropriate package
depending on Qt version selection.

Recent openscenegraph-qt provide a osgQt/Version header which contains the Qt
version it was compiled with. If this is not present, assume it was Qt4. This
works well enough in practice on the supported architectures.

The check is done with an obfuscated sed program, that outputs 0 in case of
version mismatch (and nothing otherwise) to make sure that this dependency does
not match the required ABI (or that the test does not otherwise influence the
regular osg-qt ABI check).