[mk/depends] Better handle PKGREQD containing multiple alternative choices
In case PKGREQD contains multiple alternative choices, like {alt1,alt2}-pkg,
don't warn it's ambiguous but rather restrict the list of checked alternatives
to what was found, so that the final PKGNAME is one of those.

This is just based on string matching and it assumes an or list. This does not
parse the {,} syntax, but it should be good enough in practice. PKGREQD will be
properly tested against the final PKGNAME in the cbbh rules anyway.

While here, do the same when a single alternative choice is requested in
PKGREQD. This greatly simplifies this code.