[mk/bulk] Search for "not available" packages in PKG_PATH as well
When a dependency is not found, it is checked if it is found as "not available"
in the local bulk tree. This propagates the "sysdep missing" error to dependent
packages.

This check is now extended to paths found in PKG_PATH as well. This is useful
for bulk builds of the wip/ subdirectory, when a package depends on a robotpkg
dependency that could not be built because of missing system dependencies. This
prevents from reporting the package in wip/ as artifically broken and reports a
proper "N/A" instead, with typically the missing sysdeps et al.