[mk/internal] Fix glitches in dependency reduction with alternations
Fix reducing e.g. "x>=1" + "{x,y}>=2" to "x>=2"
This was reducing to "x>=1 y>=2" before, which is wrong.

The fix consists in ignoring the branches in the alternation expansion that
cannot be reduced (the "x>=1" + "y>=2" -> "x>=1 y>=2" in the example above)
and keeping only those that reduce (the "x>=1" + "x>=2" -> "x>=2").  This
is correct because branches have the "or" semantics.

If no branch can be reduced, the alternation is left unchanged and passed to
the final merge as is (e.g. "x>=1" + "{y,z}>=2" result in
"x>=1 {y,z}>=2").