[mk/bulk] Fix extraneous space in a '--' separator
The file created in the NotAvail subdir of packages repository separates with a
"--" the names of not available dependencies from the raw error text.

However, an extraneous space was preventing this to work ("--" != "-- "), and
the db to html report tool was as a result trying to parse random text as
package names.

This has not triggered any visible error by chance (and thanks to various other
protections) until now (commit 3218775 that allows wip/ packages to pull the
"Not Available" reason from main robotpkg packages).