[mk/sysdep] Fix Qt version detection in osg-Qt
As disscussed in issue #186, the "osgQt/Version" header was recently added but
also removed shortly after. Let's not use this for the Qt version detection.

Recent osg-Qt (>=3.5.7) provide instead a openscenegraph-osgQt5.pc and
libosgQt5.so for osg-qt compiled with Qt5. Use this pattern to find the right
osgQt according to the Qt version selection. Assume Qt4 if a
bare openscenegraph-osgQt.pc exists, this is good enough for all supported
distributions.

OK Guilhem.