[mk/sysdep] In python.mk, drop PYTHON_SOABI and use PYTHON_EXT_SUFFIX instead
It seems that the python config_var 'EXT_SUFFIX' is more robust than 'SOABI' to
get the python extension name. On NetBSD in particular, only the latter is
useful (SOABI is not used on this system).

However, 'EXT_SUFFIX' is not defined in python 2.7 (it returns 'None'). So
the 'SO' var is used instead: it seems to be defined accross all python
versions.

The few PLISTs affected by the change are updated:
 - image/opencv3
 - middleware/omniORBpy