[middleware/py-omniORBpy] Rename middleware/omniORB into py-omniORBpy
This package installs the runtime python code for the omniORB generated python
backend. Prefixing the package with the python version makes it clear that the
what runtime python version is expected.

While here, fix the installation directory of the python generator, that must
go along with omniidl_be, and might use a different runtime python version.

Suggested by gsaurel@ in #207