[mk/sysdep] Avoid recompiling unrelated python files
export PYTHONDONTWRITEBYTECODE=1 in python.mk, to prevent accidental
compilation of python files that don't belong to the current package.

This solves a corner case issue where omniORB is compiled with python2 and
py-omniORBpy with python3. In this case, any omniidl invocation will compile
the python.py backend as a python2 package, while robotpkg will have compiled
it for python3. This is harnless, but it leave orphaned files appearing as
leftovers from PLIST.