[sysutils/console-bridge] Fix an issue with system package on ubuntu 14.04
On ubuntu 14.04, the version is 0.2.3, so too old for urdfdom. But the system
pkg-config file does not contain the version, so robotpkg can't check it and
uses it as is in this case.

Since ubuntu 14.04 is obsolete and soon unsupported anyway, just prefer
the robotpkg version to not bother too much with this stupid issue.