[simulation/bullet] Remove obsolete pacakge - turned into a sysdep only
The bullet-0.81 distfile has been removed from upstream, as upstream moved to
githup and kept only achives up to 0.83 (June 2015).

There is no need to have this package in robotpkg anymore, as it is available
everywhere. The default was a sysdep package always, anyway.

Move to mk/sysdep/bullet.mk