[mk/sysdep] Make sure PYTHON_EXT_SUFFIX is not referenced too early
PYTHON_EXT_SUFFIX needs a python interpreter to be computed properly, so it
cannot be used before the dependency resolution. In particular, it cannot be
used in SYSTEM_SEARCH.pkg (#212).

A possible solution could be to assign a default value, like '*.so', until the
dependency resolution has completed. This would fix #212 but it is a big magic
and the default value might not fit all contexts well.

This commit instead makes sure that an explicit error is raised if the variable
is used before it can be correctly defined, so that there are no surprise
regarding its value.