[middleware/mavlink1] Import mavlink-1.0.12 (from wip/)
This is imported as mavlink1, because the original mavlink package in wip was
versionned as 2015, so it would always be preferred over the regular 1.0.12
version.

mavlink is a library for lightweight communication between Micro Air Vehicles
(swarm) and/or ground control stations. It serializes C-structs for serial
channels and can be used with any type of radio modem.

Messages definitions are created in XML, and then converted into C header
files.