[middleware/omniORB] Install a default configuration file
omniORB provides a sample configuration file that is installed in
share/example/omniORB. Also install it in etc/ for user convenience, where
omniORB looks by default.

In order to not overwrite or delete a user-modified file, do this via a
POST-INSTALL / DEINSTALL script that checks that the file is identical to the
sample one. If not, it is left in place during removal, and not overwritten
during install.

Also add patch-ag, adding a useful option to update default configuration
file.

Should fix #215
Bump PKGREVISION