[mk] Add an intermediate variable PKGCONFLICTS and use it in pkg creation
A new internal variable PKGCONFLICTS is added. It contains the CONFLICTS
variable contents, possibly substituted by patterns defined in
CONFLICTS_SUBST.

CONFLICTS_SUBST may contain a list of substitution in the from A=B, meaning
that occurences of the string A in CONFLICTS should be replaced by B.

No functional change if CONFLICTS_SUBST is not defined.