[interfaces/humanoid-nav-msgs] Rename ros-humanoid-msgs-0.3.0 and drop ROS
This package has been provided by ROS until kinetic version. It is now not part
of ROS anymore, but still available.

In order to satisfy the few packages that depend on it and enable those package
on recent ROS version, this package is imported as a standalone pacakge (not
pretending to be ROS). However, on distributions that provide system ROS
packages <= kinetic, this dependency is set to be a ros system package.

This addresses #227.