[mk/sets] Fix pkg set expansion with patterns requiring option exclusion
When adding package patterns containing simple options requirements (or
exclusion), keep the original option string instead of replacing it with the
matched package(s). This prevents the deps reduction to merge things that have
been explicitly asked for.

For instance, the following set "*/*:*~foo */*:*~!foo" is expected to
expand to packages with the foo option and without. Previsouly, the two names
would have been wrongly merged in a single list with only the 'foo' option.