[mk/depends] Introduce a new variable: SYSTEM_DEP.<pkg>
SYSTEM_DEP.<pkg> is similary to SYSTEM_PKG.<pkg>: it names a system package for
dependencies configured as system ones.

The difference is that while SYSTEM_PKG is informal and used only in report
messages, SYSTEM_DEP is meant to name the exact system dependency and is used
(only) by pkgrepo2deb (so far) to create binary debian packages.

SYSTEM_DEP.<pkg> makes sense only on a given distribution, so as for
SYSTEM_PKG.<pkg>, only the forms suffixed by distribution, os version etc. are
supported.
Eg. : SYSTEM_DEP.Debian.libavcodec= libavcodec-dev (>= 4:0.5)

If not set explicitly, SYSTEM_PKG.<pkg> defaults to SYSTEM_DEP.<pkg> if the
later is set.

As it is very sensitive to uncontrolled information that can vary a lot with
time, SYSTEM_DEP should probably only be used on Debian systems to solve
issues not properly handled by the automatic procedure of pkgrepo2deb.