[mk/sysdep] Update SYSTEM_SEARCH for py-sip on Fedora
On Fedora, sip.so is located within the PyQt{4,5} directory and provided by
separate packages python2-pyqt4-sip or python3-pyqt5-sip. This doesn't match
what all other distributions do.

So, for now, the SYSTEM_SEARCH is just adapated to look in PyQt* for
sip.so. This may not be the right thing, but adding a Qt dependency here
seems overkill at the moment.