[devel/ros-catkin] Fix generation of CATKIN_MARKER_FILE
Fix the post-install script so that it generates an empty marker file instead
of a one containing a \n.

This is important because the CATKIN_MARKER_FILE contents is used by ros
setup.sh in `etc/catkin/profile.d/1.ros_package_path.sh` to build a
ROS_PACKAGE_PATH.  When the marker file is not empty, it's content is used as a
list of paths. When the marker file is empty, the marker file path is used
instead.

Fixes #247.
Bump PKGREVISION.