[graphics/openscenegraph] PREFER the robotpkg version by default
Depending packages require the collada plugin. Unfortunately, no distribution
provides this in their system packages, so use a PREFER robotpkg by default.

It might be better to provide a separate package just for the collada plugin,
but it has to be assessed if a single plugin source would be compatible with
diffent system osg versions. #264.