[mk/language] Add C11 language flags
This simply requires gcc>=4.9 (clang experts need to check what is needed).

For gcc>=5, the default is already C11, so nothing is required.
For gcc-4.9, -std=c11 must be passed. To avoid surprises, -std=gnu11 is
preferred, though, because -std=c11 was triggering unexpeted behaviour
regarding posix RT support ... gcc-4.9 is pretty old anyway, so not a big
deal?