[meta-pkgs/ros-base] Fix ros-noetic definitions
Fix wrong commit:4e3d359. DEPEND_ABI for ros packages is defined by the
alternative choice, so there cannot be a default for a specific version, as it
was mistakenly done.

While here, remove the DEPEND_ABI default which in practice is useless, to
avoid making the same mistake again.