[mk/sysdep] Fix python compiled file names (revert 09f3c4d)
Not sure why 09f3c4d was introduced, but it seems wrong (6 years after ...)
At least, this breaks image/opencv4/PLIST with the config-3.8.py file name.

There does not seem to be any other python file in all existing PLISTs with
dots in the basename, so this should not hurt.