[mk/fetch] Make sure mirror-distfiles does not abort for non-existent packages
When run on a package set, mirror-distfiles may try to mirror impossible
package names, pulled by dependencies. Instead of stopping, make sure the
errors are logged so that the rest of the set can be processed.

This is done by checking the 'cbeh' target first, and running the actual
mirroring targets in a sub make.