[mk/sets] Handles pacakge sets targets from toplevel directory
The sets targets can run from any directory. But if run form a directory that
defines some alternatives on itself or for some dependencies, those will be
implicitly inherited for the whole set. In order to avoid that, change to
toplevel directory and reset any such settings before doing anything.